fix: authentication
