+++ src/main/java/com/group2/glamping/auth/JwtAuthenticationFilter.java
@@ -60,7 +60,9 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
                 );
                 SecurityContextHolder.getContext().setAuthentication(authToken);
             }
+            System.out.println("hehe");
         }
+
         filterChain.doFilter(request, response);
     }
 }
