+++ src/main/java/com/group2/glamping/auth/JwtService.java
@@ -17,8 +17,7 @@ import java.util.function.Function;
 // Binh
 public class JwtService {
 
-    private static final String SECRET_KEY = "SNMBP1D7/V2QFUC4Twr1cZgictGITTT6FSJrA3Y6o2nljOsSgrAUNrvQIFgtDKvwkTkemjEPMTa9pRZvsDhpQjuSJc9I4dNMQ/Y8O6DZdE7pWH9ISEQKpAxwU4cQ4sqUVYEnzLEU8+GFMihfE/Z/jESlKpwcaPuT2ygcQwBTz3of2M0K3yEFQK9usaKKX9+C2ryu6F5Q6rclJUd7DlBxTdRRstqCumu3EEAVSwqzXEBdx73TV3cFAxh93YpzaFjCc6xTaKZR7Sb5irKVPXH0rW0c+8Zi+KVVxh5PdTDuDCVqwZn2O6bpDNFpG1yK4g1fcsDfH2+XWXLeMMdpSkUWTTPDW0tRF+jas/ss70kuXvI=\n" +
-            "\n";
+    private static final String SECRET_KEY = "5YHYOKfGQSN/DzvZgieCfh7dMOVcej+bsEePD1CKmVPtHkFHi7sXvZH48+JbfaZXbdpxX2SCe5jjxvriUlq2eBJpojTlSUI6+sOzSh4kL5if2ZEdaBDBAauUNRu5DaxtkPncW68sLO18nTkCI4AXUct8yTOF4BaLNYP7AFDsdAvNRdcda/pFKDn7jud68+BK9ixOCEOi+ZFFGfjDEBSEVZDLJYqnZ2RmSGmh1zBNhT3JUNjXEoI6K1myN4SctQgT1JbqIFi79Cay3emlphKtw3VgYPWp8bpigCf4YrfDOhn0RfoD1Ujx+NwQIeAC3FfX319GKNpuEMGtFZZIdIoLv6VqfMUwDXUlofVV9NXyjS0=\n";
 
     public String extractUsername(String token) {
         return extractClaim(token, Claims::getSubject);
