+++ src/main/java/com/group2/glamping/model/dto/response/AuthenticationResponse.java
@@ -1,5 +1,6 @@
 package com.group2.glamping.model.dto.response;
 
+import com.group2.glamping.model.entity.User;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -14,6 +15,7 @@ public class AuthenticationResponse {
     private String accessToken;
     //private String refreshToken;
     //private String expiresIn;
+    private User user;
     private String message;
 
 }
