+++ src/main/java/com/group2/glamping/model/entity/Report.java
@@ -1,6 +1,5 @@
 package com.group2.glamping.model.entity;
 
-import com.group2.glamping.user.User;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
