+++ src/main/java/com/group2/glamping/model/enums/Role.java
@@ -1,6 +1,6 @@
 package com.group2.glamping.model.enums;
 
 public enum Role {
-    ROLE_USER, ROLE_ADMIN, ROLE_MANAGER, ROLE_STAFF
+    ROLE_USER, ADMIN, MANAGER, STAFF
 
 }
