+++ src/main/java/com/group2/glamping/repository/UserRepository.java
@@ -8,8 +8,5 @@ import java.util.Optional;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, Integer> {
-
     Optional<User> findByEmail(String email);
-
-
 }
