+++ src/main/resources/application.properties
@@ -1,8 +1,6 @@
 spring.application.name=glamping
-
-
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.datasource.url=jdbc:mysql://localhost:3306/glamping
 spring.datasource.username=root
-spring.datasource.password=Huatanthinh123
+spring.datasource.password=12345
